Move ExportState into aws.Client
